[KFI]test(controls): mocks folder rename + toJson test helper
